refactor: move code of tasks and loading to separate files
