+++ src/view/loading.js
@@ -1,0 +1,7 @@
+export {renderLoadingComponent};
+
+const renderLoadingComponent = () => {
+  return `
+    <button class="load-more" type="button">load more</button>
+  `;
+};
